refactor: decouple domain models from UI concerns
- Remove watchedAt from Film (history concern, not film property)
- Create HistoryFilm model in feature/history layer
- Migrate HistoryViewModel and HistoryFilmAdapter to use HistoryFilm
- Move OnboardingItem from domain to feature/onboarding module
- Rename image → imageResId for clarity
- Update all imports accordingly
