+++ domain/src/main/java/com/drs/auralife/domain/model/Film.kt
@@ -9,6 +9,5 @@ data class Film(
     val description: String,
     val category: String,
     val episodeCount: Int,
-    val watchedAt: Long = 0,
     val modifiedAt: Long = 0,
 )
