--- domain/src/main/java/com/drs/auralife/domain/model/OnboardingItem.kt
@@ -1,6 +1,0 @@
-package com.drs.auralife.domain.model
-
-data class OnboardingItem(
-    val image: Int,
-    val title: String,
-)
