+++ feature/history/src/main/java/com/drs/auralife/feature/history/model/HistoryFilm.kt
@@ -1,0 +1,9 @@
+package com.drs.auralife.feature.history.model
+
+data class HistoryFilm(
+    val slug: String,
+    val title: String,
+    val posterUrl: String,
+    val description: String,
+    val watchedAt: Long,
+)
