+++ feature/onboarding/src/main/java/com/drs/auralife/feature/onboarding/adapter/OnboardingAdapter.kt
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.drs.auralife.domain.model.OnboardingItem
+import com.drs.auralife.feature.onboarding.model.OnboardingItem
 import com.drs.auralife.feature.onboarding.R
 
 class OnboardingAdapter(
@@ -35,7 +35,7 @@ class OnboardingAdapter(
         holder: OnboardingViewHolder,
         position: Int,
     ) {
-        holder.imageView.setImageResource(items[position].image)
+        holder.imageView.setImageResource(items[position].imageResId)
         holder.textView.text = items[position].title
     }
 }
