+++ feature/onboarding/src/main/java/com/drs/auralife/feature/onboarding/model/OnboardingItem.kt
@@ -1,0 +1,6 @@
+package com.drs.auralife.feature.onboarding.model
+
+data class OnboardingItem(
+    val imageResId: Int,
+    val title: String,
+)
